pyproject.toml
[project]
name = "ds-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "metaflow>=2.19.16",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
